requirements.txt
scikit-learn
python-dotenv
numpy
pymongo
certifi
kaggle
pandas
pyyaml
imblearn
xgboost
mlflow
fastapi
uvicorn
kafka
kafka-python
#-e .